Refactored navbar to its own component
